+++ server/package.json
@@ -1,9 +1,9 @@
 {
   "name": "learn-express",
   "version": "1.0.0",
-  "main": "./index.js",
+  "main": "./src/index.js",
   "scripts": {
-    "start": "node index.js"
+    "start": "nodemon index.js"
   },
   "keywords": [],
   "author": "Fabrice Umuhire",
@@ -24,5 +24,8 @@
     "multer": "^1.4.2",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0"
+  },
+  "devDependencies": {
+    "jest": "^26.4.2"
   }
 }
